requirements.txt
# web server with Flask and Bootstrap4
pyFlaskBootstrap4>=0.2.23
flask-dropzone
# https://github.com/SmartDataAnalytics/OpenResearch
ConferenceCorpus>=0.0.26
# https://pypi.org/project/wikirender/
wikirender>=0.0.34
# https://pypi.org/project/pylodstorage/
pylodstorage>=0.0.69
lxml #ToDo: seems to be missing in one of the imported modules
bs4
#https://pypi.org/project/odfpy/
odfpy
# http://pydispatcher.sourceforge.net/
PyDispatcher
# https://pypi.org/project/APScheduler/
APScheduler
XlsxWriter
# https://openpyxl.readthedocs.io/en/stable/
openpyxl
# https://github.com/WolfgangFahl/pyOnlineSpreadSheetEditing
pyOnlineSpreadSheetEditing

requests~=2.26.0
Flask~=2.0.2
python-dateutil~=2.8.2
tabulate~=0.8.9
WTForms~=2.3.3
pandas~=1.3.4
setuptools~=57.0.0